Move lpp.lib.plotting to a new module lpp.plotting
